agent_orchestration_multiple_tool/requirements.txt
requests==2.32.5
pydantic==2.12.2
loguru
python-dotenv==1.1.1
